+++ test/unit_tests/test_sns.py
@@ -13,13 +13,11 @@ def test_sns():
     """
     template = Template()
     title = 'test'
-    topic_name = 'test_topic'
     display_name = 'test_display_name'
 
-    sns_topic = SNS(unit_title=title, template=template, topic_name=topic_name, display_name=display_name)
+    sns_topic = SNS(unit_title=title, template=template, display_name=display_name)
 
     assert_equals(sns_topic.sns_topic.title, (title + 'sns').lower())
-    assert_equals(sns_topic.sns_topic.TopicName, topic_name)
     assert_equals(sns_topic.sns_topic.DisplayName, display_name)
 
     assert_equals(len(template.outputs), 1)